+++ mods/value/pricing.lua
@@ -25,6 +25,44 @@ function M.getSystemDescription(id)
 end
 
 do
+  local ITEM_RARITY_POOR = 0
+  local ITEM_RARITY_COMMON = 1
+  local ITEM_RARITY_UNCOMMON = 2
+  local ITEM_RARITY_RARE = 3
+  
+  local gsub = string.gsub
+  
+  local function getItemId(code)
+    return gsub(code, "^(%d+).*", "%1")
+  end
+  
+  local GetItemInfo = _G.GetItemInfo
+  local threshold = ITEM_RARITY_POOR
+  local cache = {}
+  
+  -- Check if an item is worth valuing beyond default price (vendor)
+  -- - internally checks by minimum rarity
+  -- - TODO(@fondlez): add UI option for minimum rarity for market value
+  function hasMarketValue(code)
+    local id = getItemId(code)    
+    local rarity = cache[id]
+    local _
+    
+    if not rarity then
+      _, _, rarity = GetItemInfo(id)
+      if not rarity then 
+        A.debug("[PRICING] No rarity for this item. Code: '%s'", tostring(code))
+        return true -- be conservative, no filter on lack of information
+      end
+      
+      cache[id] = rarity
+    end
+    
+    return rarity > threshold
+  end
+end
+
+do
   local makeItemCode = util.makeItemCode
   
   function M.value(token)
@@ -33,10 +71,12 @@ do
     local code = makeItemCode(token)
     if not code then return 0, NONE end
     
-    local index = systems[db.pricing]
-    local value = systems[index].func(code)
-    if value and value > 0 then
-      return value, db.pricing
+    if hasMarketValue(code) then
+      local index = systems[db.pricing]
+      local value = systems[index].func(code)
+      if value and value > 0 then
+        return value, db.pricing
+      end
     end
     
     index = systems[defaults.pricing]